Change service name and service id for traces
Now that we've got distributed tracing it makes it really difficult to
filter using tempo. To this vein the model UUID is put into the service
ID and the type of jujud is placed as the service name. Filtering becomes
a lot simpler, as the resource name now matches what tempo expects.

To show this working, adding context.Context to the uniter helps to
trace the executor. Adding a trace to the resolver is a mindfield.
We should be very careful about adding a trace to high level, as a
trace can be left running for a very long time (days+).