Add context.Context to operations
Operations are the pieces of the uniter that perform the operations.
Passing the context will then allow pushing the trace to the lower
levels and then even the run hooks. It wouldn't be out of the relm
of possibilities to trace a hook call.

This wraps up nicely how to push the context through the uniter.