Objectstore service
The domain objectstore service starts to expose the requirements
that the objectstore will require for turning a key (juju path) to
a persisted file store path. The two __may__ be the same for local
paths, but for remote s3 backed object stores, they will be
different. This mapping allows this from a highlevel.

The methods we need are simple, Get, Put and Remove. In addition, we
need the ability to watch when a new one is added and then list all
of them, so we can ensure we have the full list.