Wire up the env hooks with tracer
Adds the tracer to the env hooks. Pass over the trace and span ids
and flags will have to be done later once PR#16575 has landed.

This facilitates the ability to trace from juju unit agents though
the charm code. By just adding the JUJU_TRACE_ID and JUJU_SPAN_ID
any language or framework can pick those up and work on a trace
using that as the parent.

Most of the code is just pushing context.Context everwhere.