Refactor of DefaultAttributeValue.
In previous work around ModelDefaults we took the approach of modeling
DefaultAttributeValue to include the 3 possible sources of defaults.
After investigating the code some more we have a lot more places where
default values come from that are acting like model defaults but are not
described as such.

Think of these extra sources as ad hoc places in the code base where we
have decided to start acting with opinionated defaults about model config
and skipped the established patterns around how model defaults work.

In order to move Juju over to the services defined in both modeldefaults
and modelconfig we need to gather all of these opinions and sources of
defaults into one location and represent the values in one coherent way.

This change tackles the above problems by defining a single default
value, the source of the value and a way for the source to inject a set
of opinions on how the default value should be applied.

Overall it would be better if we could make model defaults much more
coherent and standardised but for the time being this allows use to take
what we already have move across to DQlite and the services layer.