Object store domain state layer
Implementation of object store domain state layer. The concept of
the state layer ensures that you can have multiple paths pointing
to the same hash. As long as the metadata hash and size are the same
on insertion then you can add them.

As we're going to be storing the hash as the file path for some of
the backend, they need to be unqiue. It's possible we could use
a alias instead of the hash to make it readable, but we can do that
later if required.

I had to change the schema to push the size into the metadata as it's
possible, but unlikely that a size could be different for a hash. In
that scenario the hash function is broken and we can't accept the
metadata.