Adds a model defaults table to the controller.
With this commit we are introducing a new global table to the controller
database to describe model defaults that should be applied to all models
in the controller.

Previously in Juju we have not had this notion of controller wide model
defaults. They have always been scoped to the cloud or cloud region.
However secretly under the covers we have been living a life of shame
where we have been taking the controller models config and using that as
a controller wide default source.

It is time to rise up out of the shadows of this murky controller model
defaults world and own our desire for a controller wide model defaults
source. No longer should lesser models live in shame and more
importantly no longer should the controller model be treated as some
special case model that acts as a source of truth.