Wire up objectstore to the service factory
Interestingly we can't have two objectstores called objectstore,
because we smoosh the model and controller service factories to give
you one. This might have been a mistake. Instead it might have been
better to just ask for the Controller() and Model() one. I'll ponder
if we should change this.

For now I've just called the controller one differently to prevent
the collision.