Implement the fileobjectstore
Implementation of the file objectstore. The file objectstore is a
content addressable file store. The path can be viewed as the key
to retrieve the content on disk. In reality, the key is associated
with a hash and a size, so each file requires a lookup to the
metadata to retrieve the hash. This setup allows us to deduplicate
files that have the same hash, which will manifest itself via
charms. As charms often have the same binary blob for the same base
then storing only one item instead of two is essential. It doubles
down when you deploy the same charm but with a different name, as
that also becomes deduplicated.

The code itself is written so that all requests are sent to a
central loop. This pattern gives us some advantages:

 1. We can sequence get requests after put or remove requests.
 2. When we move to HA, we can ensure that we request other blobs
    from other controller nodes before get requests. Giving us a
    sequencing advantage.

Adding HA will be done as a later addition, but can be sequenced
in the loop.

The code does pay attention to not remove a file if there is an
existing reference to the file. In some scenarios manual intervention
will be required to remove the file. This could be later done with
a worker to against all models to prune any oprhaned hashes. As we
know from the model that if a hash is not referenced anywhere we
can remove it.