Support for setting CloudDefaults.
This commit adds support set cloud defaults and overriding and
previously cloud defaults that are on a cloud. This work is to help
support future bootstrap operations.

Adjusts the state CloudDefaults() func so that when requesting cloud
defaults for a cloud that doesn't exist you get back a NotFound error.