Validate api client imports
To ensure that we don't accidently bring in an import for the api
client packages, I've added a static analysis to validate that.

In PR[1] it was noticed that we brought in the whole of the aws
library. That wasn't intentional, so we need to prevent ourselves
from doing it again. This should hopefully go someway to do that.

 1. https://github.com/juju/juju/pull/16633