Upgrade to the latest OTEL
Unfortunately OTEL decided to use an embedded interface with a lower
case method to ensure that you only used their spans. This is silly.
I've migrated from Jaeger to opentracing and now to OTEL. Each with
a different set of interfaces. Instead we have our own managed span
that abstracts this sillyness away. The solution was to just copy
their span interface minus the embedded interface and return that
from a shim type.