Test the bootstrap worker
Scaffolding the bootstrap worker before implementing the various
lift and shift code from cmd/agent/bootstrap. This should ensure
that we can deploy a controller charm and populate the agent binaries
correctly.

There is still a lot of work to be done after this one, but the
concept is sane.