Remove bootstrap flags argument
It has always been possible to pass a bootstrap flag as an option
to bootstrap-state, except that it's never different than what it
is currently. More importantly, the bootstrap worker needs to read
in the file and having it changable complicates the code. Having
a static file name means that we just read the file name from the
data dir and be done with it. Attempting to second guess it makes
the code a lot more complicated.