Implement agent binary strategy
We only ever need to upload the agent binary once in the workflow.
This allows us to do that by calling out to the binary uploader.
The worker will prevent any other dependency engine entities starting
and thus we're at liberty to perform the required steps to seed the
controller.