Removes the bootstrap prepareTools
Moves the bootstrap prepareTools that was previously run by the
cmd/agent/bootstrap state command, to the bootstrap worker. The
strategy is actually in the internal package, which gives us
freedom to implement the strategy based on the dependencies. The
uploading of the tools only happens on in IAAS models. There is no
if isCAAS check, it's up to the dependency engine to give us the right
dependencies. The code path becomes branchless on that dependency.

The userdataconfig removed the tools binary after it ran, this was
problematic as the worker isn't guaranteed to have run with in that
window. Removing that code and pushing it into the strategy now ensures
it correctly cleans up it's own concern. If the upload fails, the
tools file will still be there, allowing us to retry later.

Lastly, the big major change is to run the bootstrap worker only if
the bootstrap-params file is available. Only once the bootstrap worker
has completed do we remove that file. This should allow us to perform
retries at a later time if we want to make bootstrapping more reliable.