Specifically notify when actions fail
Since Juju 3.0, juju run by default queues commands and then waits until
they're complete and shows the results in a tabular format. But Juju does
not show the status of actions (completed, failed) in this tabular output
for singular actions, so if an action fails, this is often not communicated.

When an action fails, it fails with a message. Print this message in the
tabular output alongside any set results

Note: changes are only required for singular action runs. If an action
is run on multiple units at once, the 'plain' output actually prints the
entire yaml

Resolves:
https://bugs.launchpad.net/juju/+bug/2037279