Rework the logger in various packages
The logger in the charmhub and charms facade packages require a
cyclic dependency on the interface. This causes lots of problems
when impelementing the logger. To make matters worse, the interface
also had a loggo.Logger attached to some loggers, which makes it
impossible to satisfy that interface if you've got a sub-type of that
interface.

The solution is to make two interfaces a logger factory and a logger.
This removes the cyclic nature of the interface and cleans up what
the actual dependency is.