Ensure that we clean up the params file is CAAS
Luckily this was picked up in the github PR tests, but shows directly
that manifold config just is useless when it comes to strict typing.
The solution is to use field less structs, so any change to the
underlying struct is an error. Except that the readability dramatically
reduces. Of all the things in go, this is the worst part of it. I'd
love to have a vet or keyword that ensure all fields are added for
certain types.