Flag service
The flag service allows setting boolean values assoicated to a name
that will persist in the database. The original purpose to this, is
to be able to register if the controller has been bootstrapped. We
can envisage more flags being added over time, hence why making it
into a real service.