Add MachineRef to state
When a method in state takes a object from state as a parameter
to that method, it is anti-pattern. Generally, speaking the method
wants to reach into the other types private fields to access them
without any limitations.

Looking at AssignToMachine, the actual field members this requires
is actually smaller than the Machine doc itself. Fixing this with
AssignToMachineRef allows the sub-selection of fields that are
actually required.

Interestingly, this could be a path forward to migrating to dqlite
without ripping up the world for complex types. Although the work
to do that, requires some upfront effort.