Refactor state out of apiserver
For uploading a charm it's required that you also have the ability
to prepare and upload to the legacy state package. As we're moving
away from that package, we need to start lifting it up higher through
the callstack. This change does that, so that I can then replace it
at will later on.