Add retry logic to lxd image download
We have been having trouble with faulty service from ubuntu cloud image,
leading to bootstrap failures. In this case, we should try
re-downloading the image a few times, as we may be successfuly 2nd or
3rd time

Subtle change to contexts in bootstrap. Placing stdCtx into the cloud
call context ensures that if ctrl-C is pressed, this is transmited to
the providers via the providerCallContext.

There is a lot of bad stuff here. This is an incremental improvement.