Objects HTTP get request
Adds the ability to get an object of any type just by the path of
the object. Currently this is the storage path to the metadata, but
conceivably be the uuid of the object.

This piggy backs off the current charm objects implementation. This
is a raw endpoint and doesn't require checking of the underlying
entity state. It's either in the object store or it isn't.

---

Interestingly, I now have a better understanding of how to fix the
API server and remove a lot of complexity. The API Server should
offer a endpoint register from the workers and each handler can
then register and unregister at will. That way you can move a lot
of this mess out of the apiserver package into individual worker
handlers. Cleaning up the entire mess.