Add upsert method in state for subnets domain
This patch adds a UpsertSubnets method, which takes a list of subnets to
add or update (if they already exist).

This method is necessary for the upcoming space service layer, which has
the method SaveProviderSubnets. This method is currently not atomic on
our current mongodb state, but the idea is to turn it atomic by running
all the subnet creations/updates under one single transaction. Thus, the
UpsertSubnets method will be used.