Wire up bootstrap deployer from the worker
Now that the bootstrap deployer has all the relavent data and tests,
this commit wires up the actual dependencies to ensure that we can
drive the bootstrap package.

Tests for the bootstrap worker will land following this patch. Most
of the work here is to ensure that we can correctly bootstrap and
also that we can deploy the controller charm.

Changes of note, are that we have a various interfaces that split
the logic into descrete "services" that can be fulfilled with
the domain services packages once we implement applications, charms,
machines and units. For now they're still backed by state.State.

The only ugly aspect of this, is the logging interfaces, but
that will change once we move to slog.