Add network domain service, SaveProviderSubnets method
This patch creates the new network domain service layer, by adding the
first method SaveProviderSubnets.

It must be noted that previously, on the legacy mongodb state, this was
not done atomically: it was done by updating/adding subnets one after
the oter not within any transaction or locking. This may have lead to
races. This method now makes use of UpsertSubnets which takes a slice of
SubnetInfo and perform all the operations within a single transaction.