This PR implements support for waiting for Pebble Notices and firing them as events to the charm. Pebble only has one notice type implemented right now, the `custom` notice, so that's all we care about for now. This gets sent to the charm as a `<container>-pebble-custom-notice` event. In this event context, the following environment variables are set:
* `JUJU_NOTICE_ID`: the unique ID of the notice
* `JUJU_NOTICE_TYPE`: the notice type (currently only `custom`)
* `JUJU_NOTICE_KEY`: the notice key, for example `example.com/foo`

The change adds a new worker, `pebbleNoticer`, which is similar to the
existing `pebblePoller` worker that handles
`<container>-pebble-ready`. It uses the Pebble client function
`WaitNotices` to wait for new notices to arrive, and when custom
notices arrive, it sends them to the charm using the existing
`WorkloadEvents` abstraction. There's a tomb for the worker, and one
`run` goroutine per container.