Propagates errors from start instance for each zone.
When attempting to start instances for a slice of zones it can be the
case that the cloud provider fails to start the instance in any of the
availability zones due to an account issue such as running out of
resources.

What we were not doing is transmitting the errors from start instance
back to the caller in this case and so we would end up with an error
that contains no information as to the actual problem.