Fix flaky block device tests
To improve the robustness of our tests we must use the TxnRunner,
which then gives us a retry mechanism. For simple unit tests we can
get away with just using .DB() method on the test suites. For
integration tests that's not possible.

In addition to the flaky tests I've also fixed the argument passing
in the tests. You can't use fmt.Sprintf because it's still possible
to bypass quoting rules.