Homogenize vm naming in azure and aws
VM and EC2 instance names in Azure and AWS repectively were not
consistent with juju machines naming on other providers. In all other
providers the naming is `juju-{namespace-short-sha}-{machineID}` (e.g.
juju-abc1234-0) while on Azure it was `machine-{machineID}` and in AWS
it was `juju-{machine-tag}-{namespace-short-sha}`.
Now they all follow the same pattern.