Prevent file backed objectstore in ha mode
We don't need to provide file backed object storage in HA mode.
Instead we want users to use a third party service (minio or similar)
for those requests.

The code is really simple, when moving into enable-ha, if it's backed
by file, just present an error that it's not possible.