Expose currently unused RootHTTPClient for S3
RootHTTPClient returns an HTTP client that sends requests to the apiserver
root address. This is and will be most useful for the s3 endpoint, whose
specification assumes the first path location is a bucket.

Use this client in the s3client in internal, slightly simplifying + generalising
the client construction and options

Also regenerate mocks and expand stubs. It turns out in a bunch of places we
generate mocks for APICaller