Implement a exclusive lock claimer
The claimer is based on the leases subsystem. We can use the leases
to provide an exclusive lock to prevent us from adding or removing
a file if there is a worker cleaning up files in the background.

The base objectstore with lock requires some tests, but the rest is
in place.