Add S3 Put endpoint for local charms
Add new endpoint which (for the time being) shims around the already
existing charm uploader handler.

We need this new endpoint because the current endpoint breaks down
the charm url into it's component parts. However, we're moving towards
treating the charm url as an opaque string before removing it entirely,
so we can drop series from our codebase.

Implement a new putter in the client to speak to this endpoint.

For compatibility, construct a fallback putter, which iterates through
putters until one succeeds. For this, we need to be able to parse
non-JSON output from the webserver underlying Juju. This is where the
changes to api/httpclient come from