Extract secretary finder from worker
The secretary finder is used to ensure that lease claims are valid
before querying the manager. These are simple checks; the lease
holder is a valid tag, or the lease key is a valid application tag.

Initially those where in state, but then moved to the worker.
Except that's also the wrong place. The lease worker shouldn't have
any knowledge of an application or controller UUID. With the
introduction of an objectstore lease, the worker will now know
about those. Instead we lift the secretary finder out of the worker
and into the machine manifold. The machine manifold is the location
for all dependencies to be tied together. This is the location to
register the indvidual secretaries.

The objectstore secretary will be added in the next commit.