Ensure goroutines do not leak
Now that we've added goroutine leak detection, we can start using it
in test suites. This should ensure that we're correctly cleaning up
workers when we kill them.

This did catch a lot of tests which were leaking, so it's nice to
clean up those directly. This should ensure that our tests don't
become bogged down and repeated runs are possible.