Add permissions schema.
Permissions are related to user for a variety of things: models,
controllers, appliation offers, and clouds.

There is a typed list of access values both in juju and now in the
schema.

The combintion of the user id and the object key must be unique within
the table. This helps ensure that we do not have competing references to
a user's access.

The user should exist before a permission is added for it.