Prevent segfault when closing dqlite app twice
If calling close on the dqlite app multiple times, it will cause
a segfault. Segfaults are non recoverable. To prevent any mistakes
in the future, this makes close on the application idempotent. It
will return the same error for every time you call it.

This fixes a long standing issue in some of our tests which ensure
that we don't leak any goroutines during rounds of testing. We've
yet to see this in production as the code flow is serialized and less
likely to happen.