ObjectStore: S3 Static Credentials
Enables S3 static credentials to be added and modified on the
controller config. In an ideal world we'd base the s3 credentials
on the secrets work we have. Unfortunately we're currently not
there. There is a chicken and egg situation for how to grant the
use of a secret. Whilst we work out a valid use of STS (security
token service) and temporary credentials, and also not storing them
in them in similar places as the controller config, the following
PR adds legitimate static credentials. The use of these static
credentials shouldn't be used in production without understanding
the consequences.

One possible solution could be to use file based object store, add
a relation to vault in the controller charm, which can then add
them to a databag (leader only). Thus allowing the controller to
then use those to enable s3 and HA (draining from file to s3 as
part of the s3 enablement). This is a bit of a workflow and might
not work in practice.