Use RootHTTPClient in migrationtarget client
All the apiserver endpoints for migration are scoped to the server root
(i.e /migrate/blah instead of /model/:modeluuid/migrate/balh). See
apiserver/apiserver.go

However, the migrationtarget client uses an http client which attempts
to prefix urls with "/model/:modeluuid". It seems this fails and we rely
on this failing

This works, but is confusing and potentially delicate. I suspect this
was only done because this was the only HTTPClient available at the
time.

Use the newly made available RootHTTPClient instead to avoid this
misdirection