Adds controller system ssh key to controller config.
Every Juju controller has generated for it a system ssh key that allows
the controller it self to ssh into machines and containers.
Prodominently this has been used for configuring machine based
firewalls.

A fixed was introduced as part of #16394 that took the juju-system-ssh
key as part of the controllers model config and appended it to newly
created model's authorised key.

The problems this introduces is that it special cases the controller
model as a defaults source for models which is not the design Juju is
attempting to achieve. It also means that during model migrations we are
migrating the system ssh key of one controller to another potentially
introducing a security bug into the model where the previous controller
could still access machines of a migrated model.

With this PR we have mode the bootstrap client instead generate the
system ssh key and place it into the controllers config. In the API
facades we now form a union of the two ssh key sources before returning
them to the caller.

This work also allows for the continuation of moving model config over
to Dqlite as now we don't have to special case the controllers model
config in the design.

As part of this PR we have also updated some wording around model
defaults at bootstrap time that was confusing and not correct.

Newly created instance configs will also get the controllers system ssh
key.