Add tests for the objectstore worker
The worker tests now ensure that we refresh the client any time
the controller config changes. It becomes tricky, because we don't
want all the requests to go via the loop, as that would cause a lot
of contention with lots of models.

In addition, I've split out the ability to swap out a concrete
service factory type for a mock one in the manifolds. The dependency
takes a functor (A -> B) and we can just supply our own at testing
via the manifold config.