Add environ to bootstrap worker
This patch is a preamble to the work being done to remove usages of the
space/subnet domain from the legacy state. We need the environ because
in order to init the API host port at bootstrap we'll need to
retrieve the instance address from the environ in the bootstrap worker.

It must be noted that this work tends to move some logic from agent
bootstrap to the bootstrap worker.