Removes the all/multiwatcher
Removes the all/multi watcher from state and apiserver. The prior
all/multiwatcher was tied to the state documents directly. The way
we want to move forward is to pass discrete events driven by the
changestream.

There are some unfortunate consequences to ripping this out, but
probably should be tackled soon:

 1. Some bundles with unit/machine placeholders can't be deployed.
 2. The waitfor command used the all watcher api directly.

Once we have a new API for the changestream, we should bring the
waitfor command back. I'd rather not comment out code that isn't
working in any way.

There is probably a lot more dead code that we can remove as well.