Create corresponding migrate endpoint for charms
We recently added an S3-compatible endpoint to upload local charms to
replace/deprecate the old charms handler.

This is working towards the eventual aim of treating charm url as an
opaque string, so we can remove series from it. In this respect, the
important chance is that the api interface now treats charm url as an
opaque string (in a header), even if (at the moment) we parse it at the
other end

The /migrate/charms endpoint also uses this handler. So, as before,
create a new endpoint for migrating charms which uses the new handler.
This takes us a step closer to deprectaing this handler

As a consequence, re-write the controller api to upload charms during
migrations to use the new endpoint.

NOTE since we can only migrate forward versions, we can drop the old
UploadCharm api client method as we do not need backwards compatibility

This means we drop another instance of reading Series from charm.URL,
and even go as far as treating the charm url as an opaque string