Allow model migrations to use the model DB
This refactors the facades to hand out model exporters and importers.
It hides the controller and model DB, preventing the pollution of
very low level types to facades. This should prevent any facade
code from interacting with a database directly. We're explicitly
calling out that they should either use a service or a model migration
factory.

This was a long held design issue that we've always wanted to clean
up. I believe this is probably the best approach; it provides
better guard rails.