Ensure we import models
The migration target tests rely on the importing logic of the
underlying state. To work around this for now, we can just pretend
to import the new state, whislt passing through to the legacy
import logic.

As more logic is moved over, it might be wise to re-write these
tests again.