Event Watcher
The event watcher dispatches events for the apiserver watch methods.
This is to replace the multiwatcher setup. Instead the idea there
will be one event watcher for every db. The reason for all of this
machinary is to prevent adding back pressure to the underlying event
stream. The event watcher will dispatch all the events asynchronously.
It will then be upto the api to consume the events at their leasure.

The major difference between this and the multiwatcher, is that there
is no backing store. You only get events from when you subscribed and
in the future. If the api disconnects and reconnects, they'll get
only the new events, no historicals.