Remove legacy fallback case
This case, where we fallback to the series in a charm's url, is a
fallback case for charmstore charms from before charm series was
included in charm metadata

We stopped supporting such charms a long time ago. We will failure
earlier for such charms.

Remove this case. As a side effect, we remove a reference to the Series
attribute in of charm URL