Introduces agent version to dqlite models.
With this commit we are adding a new model agent table that will drive
the information around what agents are running for a model. Specificly
at the moment we are concerned with describing the agent version that
model should be running.

Previously this has been kept in model config as a special case read only
attribute that has a lot of logic about when it can and can't be changed.

By making it a first class citizen of a models metadata we can treat it
properly and remove a lot of complexity out of modelconfig. It also
allows us to not use the controllers model as a special case for sources
of default information.

This is a placeholder commit to introduce the table into the schema.