Adds setting and configuring agent version.
During model creation and bootstrap we are adding support for both
setting a models agent version and for also implementing the buisness
logic for determining what a models agent version should be when it has
not been supplied.

This code has been copied over from controller model config creator. For
this PR we are leaving a small todo around the tools finder that will be
implemented in a follow up PR when the API facade gets wired up.

- Adds logic to model state for setting the agent version record on
  model creation.
- Adds service layer logic for choosing and validating the model agent
  version chosen based on the current rules Juju implements.
- Defines a slim downed tools finder interface for model service to
  check for available tools.
- Updates the bootstrap model process to also calculate and validate the
  models agent version.