Wires up user correlation for models in DQlite.
Until now we haven't had any strong referential integrity for model
owners onto the user table. Now that the user DDL has been established
this sets up the checks and test to assert that when adding a model the
owner exists and is a valid user in the system (not removed).

As part of this PR I have also wired up the bootstrap process to set the
admin user as the owner of the controller model.

This PR also removes some hard coded values that were living in the
wrong place. Specifically the username we give to the admin user at
bootstrap time.

Also moved controller defaults such as controller name and controller
model user to the core package so they can be referenced properly.