Move addresses and controller cloud service initialization to bootstrap worker
This patch adds two new functions to the bootstrap worker:
- initAPIHostPorts
- initControllerCloudService
For this, we needed to change the signature of SetAPIHostPorts on the
legacy state and remove the inner calls to the space state.
Note: We have commented the call to SetAPIHostPorts on the peergrouper
worker. This is temporary until we can inject the new network domain
needed to filter the management space addresses.