Remove series from charm url when adding charms in bundles
When we calculate the base to use for a charm, as well as adding the
base to the origin, we also used to convert it to a series and include
it in the charm url

However, this charm url series is never read. It was only included as a
legacy fallback option for charms which we have not supported for a long
time

Drop the series from the charm url. This is part of an effort to remove
Series from the charmurl in general, as part of work to stop using
series and to use bases instead across our codebase

As a flyby, refactor some simple inconsistant variable usage