Use correct hash: SHA384
Changes the default hashing mechanisms to use sha384 for files. This
is a change from sha256, which was used previously. Currently,
resources are verified using sha384, which means that we need to verify
the hash of that correctly. We could move all that over to sha256,
which could potentially break something, so for now I'll keep with
the default to prevent re-writing juju.

The s3 objectstore has the unfortunate issue that we need to hash
twice, once for sha384 for juju verification and once for sha256 to
verify against AWS. AWS only supports CRC32, CRC32C, SHA1 and SHA256.

I suspect long term we should just move juju over to sha256. The only
logical reason that juju might be using sha384 is to prevent potential
length extension attacks. I'm not entirely sure that attack is a
attack vector here.