Allow the use of an s3 bucket name
As bucket names are very restrictive, we need to parse them to
not cause any potential issues. It should be noted that once a
bucket name is set, it should never be changed. That includes the
URL of a bucket. If the bucket changes location then juju is SOL'd.

Most of the changes reflect the passing in of the bucket name.