Remove controller config option
There are too many foot guns assoicated with allowing a user to
pass in their own bucket name:

 1. If two controller have the same bucket name then data loss will
    happen.
 2. You shouldn't be able to change the bucket of a running
    controller. Data loss will happen.
 3. It's not currently possible without more infrastructure to know
    if the bucket name has changed.

All of this adds up to, a lot of issues for not much gain. If it is
really wanted, we can add the infrastructure to do this. Most likely
we'll have to tag buckets with a controller UUID and if it already
exists then prevent bootstrap or a worker running on restart.