Implement objectstore pruner for the objectstore directly
This reverts the prior change. We don't want to leak out the underlying
hash of the filesystem. Instead we want to contain the pruning logic
into the objectstore type. That way it can work on exactly what it
needs.

We now have a pruner that will run just the logic in the objectstore.

The objectstorepruner worker will remove the s3 information directly
using the file or s3client. We have to use the raw clients directly
because we can't access the objectstore if the model has been removed.