Reload spaces from the bootstrap worker instead of agent bootstrap
This patch moves the last dependencies on the legacy mongodb state of
the agent bootstrap regarding networking.
We move ReloadSpaces to the bootstrap worker. This required some changes
on the application (domain) of the legacy state, namely avoid retrieving
the list of spaces needed by MergeExposeSettings and pass those spaces
as parameter, thus the changes in the application facades.