Move address finding out of the worker
The worker should know specifically about what needs to be run, but
not necessarily the actual implementation details.

With this in mind, we make a BootstrapAddressFinder that we can
swap in and out depending on IAAS or CAAS.

Also standardise the function calls, using func types, with cleaning
up of suffixes around Func on functions that we're calling.