Move wait-for to expr-lang
Since implementing wait-for there have been a number of projects that
have implemented a simple expression based language. One of them is
very popular:

 - https://expr-lang.org/

The following is an exercise to see how easy it is to move the wait-for
command over to the new library.

It's extremely easy to move over to the new language. There are few
breaking changes, but I think overall it's much better to not have to
maintain it.

 1. We get very good documentation out of the box
 2. We don't have to maintain the code, where I am the only one who
    knows how it works.
 3. This isn't a juju concern.

Potentially we could use something else, but feels like the appropriate
thing to do.