Fixes minimal image selection for amd64.
Due to the way we order images from the OCI provider there was a
possability that under certain circumstances we could select a minimal
Ubuntu image for amd64.

Currently Juju does not support minimal images and only has code to stop
this from happening when the architecture was arm64.

This commit expands the catching methods to consider all minimal images
regardless of arch.