Lock object blobs with a retention lock
The retention lock allows the locking of s3 objects to prevent
accidental deletion. We have to be very careful with the mode we
use.

The documentation[1] states that "you may have an ongoing project
utilizing a dataset that you want to keep immutable until the project
is complete". This can be done via a "Legal Hold" status. Thus
turning it to ON a governance mode allows us to ensure that if
we try and delete an object then confirmation will be required.

This prevents anyone from removing an object. Juju will now
break if it can't get hold of an object blob, esp. concerning
agent binaries.

 1. https://docs.aws.amazon.com/AmazonS3/latest/userguide/object-lock.html#object-lock-best-practices